Ease of life and navigation changes
Added comments to everything for readability, used define for ENES100 fields and methods for readability, created global position variables to stop initializing location within every scope and for readability. For navigation, it now uses the desired theta coming from missionY, to do calculations, which means it should be able to navigate no matter the starting position and obstacles. Testing needed